Make SiteManagerAdapter as an interface.
