+++ log-replication/src/main/java/org/corfudb/logreplication/infrastructure/CorfuReplicationSiteManagerImpl.java
@@ -1,0 +1,50 @@
+package org.corfudb.logreplication.infrastructure;
+
+import lombok.Getter;
+import org.corfudb.logreplication.proto.LogReplicationSiteInfo.SiteConfigurationMsg;
+
+
+public class CorfuReplicationSiteManagerImpl implements CorfuReplicationSiteManagerAdapter {
+    @Getter
+    CorfuReplicationDiscoveryServiceAdapter corfuReplicationDiscoveryService;
+
+    @Getter
+    SiteConfigurationMsg siteConfigMsg;
+
+    public void setCorfuReplicationDiscoveryService(CorfuReplicationDiscoveryServiceAdapter corfuReplicationDiscoveryService) {
+        this.corfuReplicationDiscoveryService = corfuReplicationDiscoveryService;
+        start();
+    }
+
+    public synchronized SiteConfigurationMsg fetchSiteConfig() {
+        siteConfigMsg = querySiteConfig();
+        return siteConfigMsg;
+    }
+
+    /**
+     * Will be called when the site change and a new configuration is sent over
+     * @param newSiteConfigMsg
+     * @return
+     */
+    public synchronized void updateSiteConfig(SiteConfigurationMsg newSiteConfigMsg) {
+            if (newSiteConfigMsg.getSiteConfigID() > siteConfigMsg.getSiteConfigID()) {
+                siteConfigMsg = newSiteConfigMsg;
+                corfuReplicationDiscoveryService.updateSiteConfig(siteConfigMsg);
+            }
+    }
+
+    public void prepareSiteRoleChange() {
+        corfuReplicationDiscoveryService.prepareSiteRoleChange();
+    }
+
+    public int queryReplicationStatus() {
+        return corfuReplicationDiscoveryService.queryReplicationStatus();
+    }
+
+    //TODO: handle the case that querySiteConfig return an exception.
+    public SiteConfigurationMsg querySiteConfig() { return null;};
+
+    public void start() {};
+
+    public void shutdown() {};
+}
